dayly commit -GOPTaskScheduler.shortestQueueFirst use 4 parameter shortestQueueFirst(StreamGOP x,boolean useTimeEstimator,boolean virtual,double SDcoefficient) -StreamGOP(StreamGOP X) do deep clone the StreamGOP -addStream can now decide if we will merge or not for case B., need a bit more optimize than copy code to type C (for both sorted Queue and unsorted Queue) -GOPTaskScheduler.virtualQueueCheckReplace(StreamGOP Original,StreamGOP merged) return if merged request will cause any miss deadline or not. (int flag, 1=pass, 0=other miss deadline, -1=this miss deadline) TODO: try in between of unsorted list //TODO: add -2, which is both miss deadline -new Serverconfig parameter, batchqueueSorted (not configurable yet) //TODO: shortestQueueFirst also update x.estimatedExecutionSD, probably replace it to return stat datastructure. or tuple //TODO: copy lvl2 merging logic to lvl3, deduplicate code too.

+++ src/Scheduler/TimeEstimator.java
@@ -38,7 +38,7 @@ public class TimeEstimator {
 
     //design decision, should this return per segment or per action?
     //select do it all in here!
-    public static Long getHistoricProcessTime(String VMclass,StreamGOP segment,int SDcoefficient){
+    public static Long getHistoricProcessTime(String VMclass,StreamGOP segment,double SDcoefficient){
         //SDcoefficient=1 is Worst case, -1 is BestCase,
         HashMap<String,Stat> polled1=Table.get(VMclass);
         long Time=0;
